renaming IdentitySection to IdentitiesSection
